# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Quarto API documentation.",
        "title": "Quarto API",
        "termsOfService": "https://quarto.fr/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://quarto.fr/support",
            "email": "support@quarto.fr"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/challenge/my": {
            "get": {
                "description": "Get all challenges sent and received by the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Get my challenges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/challenge.ChallengeListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/challenge/respond": {
            "post": {
                "description": "Accept or decline a challenge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Respond to challenge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Response to challenge",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/challenge.RespondToChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/challenge.ChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/challenge/send": {
            "post": {
                "description": "Send a challenge to another player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Send challenge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Send challenge request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/challenge.SendChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/challenge.Challenge"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/my": {
            "get": {
                "description": "Get all games for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get my games",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Game status filter (active, finished)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/game.Game"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}": {
            "get": {
                "description": "Get a game by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.Game"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}/forfeit": {
            "post": {
                "description": "Forfeit the current game",
                "tags": [
                    "games"
                ],
                "summary": "Forfeit game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}/place-piece": {
            "post": {
                "description": "Place a piece on the board",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Place piece",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Place piece request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.PlacePieceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/game/{id}/select-piece": {
            "post": {
                "description": "Select a piece for the next move",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Select piece",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Quarto-Connect-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Select piece request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.SelectPieceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/game.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "challenge.Challenge": {
            "type": "object",
            "properties": {
                "challenged_id": {
                    "type": "integer"
                },
                "challenger_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "responded_at": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, accepted, declined, expired, cancelled",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "challenge.ChallengeListResponse": {
            "type": "object",
            "properties": {
                "received": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/challenge.Challenge"
                    }
                },
                "sent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/challenge.Challenge"
                    }
                }
            }
        },
        "challenge.ChallengeResponse": {
            "type": "object",
            "properties": {
                "challenge": {
                    "$ref": "#/definitions/challenge.Challenge"
                },
                "game": {}
            }
        },
        "challenge.RespondToChallengeRequest": {
            "type": "object",
            "required": [
                "challenge_id"
            ],
            "properties": {
                "accept": {
                    "type": "boolean"
                },
                "challenge_id": {
                    "type": "string"
                }
            }
        },
        "challenge.SendChallengeRequest": {
            "type": "object",
            "required": [
                "challenged_id"
            ],
            "properties": {
                "challenged_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "game.Game": {
            "type": "object",
            "properties": {
                "available_pieces": {
                    "description": "JSON array of piece IDs",
                    "type": "string"
                },
                "board": {
                    "description": "JSON serialized 4x4 board",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_turn": {
                    "description": "\"player1\" | \"player2\"",
                    "type": "string"
                },
                "game_phase": {
                    "description": "\"selectPiece\" | \"placePiece\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "move_history": {
                    "description": "JSON array of moves in notation",
                    "type": "string"
                },
                "player1_id": {
                    "type": "integer"
                },
                "player2_id": {
                    "type": "integer"
                },
                "selected_piece": {
                    "description": "Current piece ID to place",
                    "type": "integer"
                },
                "status": {
                    "description": "\"active\" | \"finished\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "winner": {
                    "description": "\"player1\" | \"player2\" | \"draw\"",
                    "type": "string"
                }
            }
        },
        "game.PlacePieceRequest": {
            "type": "object",
            "required": [
                "position"
            ],
            "properties": {
                "position": {
                    "type": "string"
                }
            }
        },
        "game.SelectPieceRequest": {
            "type": "object",
            "required": [
                "piece_id"
            ],
            "properties": {
                "piece_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}